refactor(data): split 1834-line market.py into providers/market subpackage
- common.py (Quote model, mock data, parse helpers), quotes.py, technical.py, financial.py, chips.py, sentiment.py, rules.py
- __init__.py keeps a facade re-exporting public + test-used symbols
- update monkeypatch targets in provider tests to new module paths
